+++ src/coupler/surface/TemperatureIndex.hh
@@ -23,6 +23,7 @@
 
 #include "pism/coupler/SurfaceModel.hh"
 #include "pism/coupler/surface/localMassBalance.hh"
+#include "pism/coupler/DebrisModel.hh"
 
 namespace pism {
 namespace surface {